Refactor: move `render` function from `handler.go` to `handler_helpers.go` for better organization and modularity.
